BEGIN_PUBLIC starlark: clarify ordering/sorting relations
Details:
- Change 'sorted' to use the decorate/sort/undecorate pattern.
  This causes the key function to be called exactly once per
  element, in order, and simplifies the error handling. Test.
- Make StarlarkList, Tuple implement Comparable and simplify STARLARK_COMPARATOR.
  (This makes assertThat(StarlarkList) an ambiguous overload, weirdly.)
- Publish EvalUtils.STARLARK_COMPARATOR as Starlark.ORDERING.
- Make it use ClassCastException instead of ComparisonException,
  as callers must already be equipped to handle it.
- Tweak error messages.
- Glob the testdata/*.sky files.

This change was split out from the discussion of floating-point numbers
in CL 335290445, after the realization that most sort algorithms require
transitivity of the ordering relation, so any ordering in which
NaN != NaN cannot be used for sorting, even though it may be desired
for the language's < operation. (Using insertion sort as a mental model
for sorting was a poor choice in hindsight because it directly compares
elements; by contrast mergesort makes inductions.)

For comparison, Python's sorts are unpredictable in the presence of NaN,
as is C++'s std::stable_sort, and Go's sort.Interface using float<float as
the order. Java's Double.compare is a total order, as is Go's sort.Float64s.
In short, regardless of what position we take on whether to use IEEE-754
semantics for float < float, sorting requires a transitive ordering.
This change enshrines that.

This change also removes the StarlarkBuiltin annotation from Sequence,
and adds cases to Starlark.type for interfaces like Sequence,
StarlarkIterable, and StarlarkCallable.
This is unfortunately a painfully breaking API change for the fragile
heuristics in Copybara's documentation-generating annotation processor.
We really need to rethink it.

END_PUBLIC

PiperOrigin-RevId: 338532782