Allow remapping of __BAZEL_EXECUTION_ROOT__ to execution root in wrapped_clang
Some clang flags requires the file path passed to it to be an absolute
path (for instance, the `-fuse-ld` flag).

Currently, in order to use those flags, we would need to either:
- Expand the location of the file before it is passed to wrapped_clang,
  which breaks reproducibility, since the absolute path would go into a
  cache key.
- Place the file at a global path (for example, `/tmp/my-linker`), and
  make that available to the build, which breaks hermeticity.

This change adds a new shared secret, `__BAZEL_EXECUTION_ROOT__`, so
that rule authors can use it as a prefix to paths that they want to be
expanded properly.