Allow Jacoco Coverage to work with classpath jar.
If the classpath for a java binary exceeds the limits specified, a
separate jar is created at runtime to contain the classpath and is the
target of execution. The JacocoRunner could not handle this case, only
inspecting the wrapping jar.

This change adds code to extract the objects in the original classpath
and pass them to the original handling code.

Fixes #11847

Closes #12325.

PiperOrigin-RevId: 338668012