Create `EventKind.FATAL` to represent a crash event.
In a followup, a `FATAL` event will be emitted during a crash. This is a new kind of event to 1) distinguish from "expected" errors (`EventKind.ERROR`) and 2) so that `UiEventHandler` knows not to show default messages ("Loading..." and "checking cached actions") after the fatal event. This way the crash is the last thing seen on the UI.

Clean up other assorted warnings and make the constant sets in `EventKind` immutable.

PiperOrigin-RevId: 338714982