Make a failure to spawn the subprocess in a sandbox exit code 36 rather than 1. It's a local environmental error for the subprocess not to start, rather than a build failure, as evidenced by the other bits: Status.EXECUTION_FAILED is a "system" error and the suggestively named LOCAL_EXEC_ERROR.
I would be happy to just migrate this code to EXECUTION_IO_EXCEPTION and change that exit code as well. This one seems more egregiously wrong, given that it has the "system" bits set in the SpawnResult, but that one also seems bad.

RELNOTES: Bazel returns exit code 36 (rather than 1) if it fails to start a subprocess in a local sandbox due to environmental issues, for example, if the argument list is too long.
PiperOrigin-RevId: 339245633