Prevent crash when using a tree artifact in cc_library's srcs
The check to determine which subset of the toolchain files to provide
as inputs requires knowing the file type. actionBuilder::getActionName()
throws IllegalStateException when given a tree artifact (directory).

In this case, we provide all toolchain files as inputs, since Bazel does
not know the content of the directory at this time.

PiperOrigin-RevId: 339271792