Add support for databinding-generated sources to depend on AndroidX instead of the support libraries.
Note that the Android databinding libraries already support AndroidX support. This CL is primarily introducing a new AndroidX databinding flag and propagating it to the necessary places. This required introducing new flags for some of the actions that correspond to kicking off various parts of the databinding process.

The new flag being introduced is: android_databinding_use_androidx.

The new tests are primarily ensuring each action tied to databinding has the new androidx parameter passed to it. Unfortunately, this only relies on rule analysis and doesn't actually verify that the actions themselves behave correctly with the flag enabled. Similarly, the databinding mapper implementation file switching logic changes can't be tested without executing the actions. Please advise on how to test this if it's valuable.

RELNOTES: Add support for using AndroidX dependencies in data-binding-enabled targets.
PiperOrigin-RevId: 339381866