Split coverage post-processing into its own spawn
This change fixes b/79147227 by introducing a flag that makes coverage post-processing happen in its own spawn. The runfiles in the coverage spawn are those of the LCOV merger instead of the test runfiles.

The flag introduced is --experimental_split_coverage_postprocessing which depends on the flag --experimental_fetch_all_coverage_outputs being enabled too.

PiperOrigin-RevId: 339699090