Use INTERNER to manage canonical instances of source artifact roots.
Blaze benchmark runs show there is no meaningful heap memory usage difference between this and the current solution, which relies on a predefined map.

In contrast, when the Interner calls are removed from this CL, Blaze benchmark consistently shows around 7MB of heap memory usage increase, which is precisely as expected.

PiperOrigin-RevId: 339959603