Remove superfluous bazel_toolchains patch
This is effectively a revert of
https://github.com/bazelbuild/bazel/commit/8946cc1746f4092d86380a23b0ac8b2d8699d3fc.

I originally needed to rename an attribute in `bazel_toolchains` for
https://github.com/bazelbuild/bazel/commit/22b4dbcaf05756d506de346728db3846da56b775. The `target_compatible_with`
attribute used in one of the repository rules clashed with the
attribute I added in the Incompatible Target Skipping patch. Once we
merged 22b4dbcaf05756d506de346728db3846da56b775 we realized that we
broke a bunch of projects because they pulled in unpatched versions of
`bazel_toolchains`.

John Cater (@katre) noticed that we can fix everything by making the
new attribute not apply to repository rules. He submitted
https://github.com/bazelbuild/bazel/commit/775a979bde21a71ddf04f67b865e8c43a29f702a to fix this.

That means that the original patch I added is actually not needed
anymore.

A follow-up patch will delete the .patch file from the source tree.

Closes #12396.

PiperOrigin-RevId: 340291830