bazel packages: delete NativeProvider, dematerialize StructImpl.provider
This change merges NativeProvider into BuiltinProvider. The two
classes were almost identical.

Also, it removes the StructImpl.provider
field to save one word of space in each instance of NativeInfo.
This requires each subclass to define getProvider,
which is straightforward but leads to a tiresome diff.
(Sadly this doesn't yield tangible space savings; I'm not sure why.)

All the interesting changes are in lib.packages. The rest is repetitive.

This is a small step towards withering StructImpl and StructApi
out of existence.

NO_IFTTT=spurious warning
PiperOrigin-RevId: 341102857