Automated rollback of commit ad181455c9bceba02cd7764d89f2f5970dee9b55.
*** Reason for rollback ***

Causes some dynamic builds to fail.

*** Original change description ***

Do not check the return value of Future.cancel()

Future.cancel() is expected to return false if the future is done or had been cancelled
previously, and there doesn't seem to be a way to check whether or not that is the case
without a race condition.

RELNOTES: None.
PiperOrigin-RevId: 341138960