Add target settings attribute to toolchain rule
This is the implementation of proposal https://docs.google.com/document/d/13yXhOQeMZQmlW3obyZ7fJP3sm3eVLbK3Nuk2LZjcZiU/edit.

The added `targetSetting` attribute, can reference `config_setting` rules. The attribute is resolved in target configuration and the returned `ConfigMatchingProvider` already has the value if this toolchain matches.

Difference to the proposal is that `targetSettings` are a `LABEL_LIST` where all the settings need to be satisfied. This might increase a bit usability. It makes the implementation a bit nicer, because instead of special handling when the value is not present, a list can uniformly handled.

Closes #12425.

RELNOTES: Toolchain rule is extended with target_settings attribute.
PiperOrigin-RevId: 341386320