Fix sleep-process not killed at exit in test-setup.sh
When sending the kill-signal to cleanupPid only
the subshell is killed but the spawned sleep
process within that shell remains and will remain
up to 10 seconds after the test has finished.

This is especially troublesome in the case of running
the test from a nfs-mount since the sleep process
may inherit file-descriptors to files within the
nfs-mount which could then not be removed until the
sleep process has exited.

Closes #12278.

PiperOrigin-RevId: 341567585