Default genquery somepath to aggregate and output in dependency order.
Somepath used in Genquery was sorted alphabetically by default which loses its previous dependency order. This CL checks for the existence of somepath at the top level of QueryExpression (e.g. `somepath(a, b)`) and changes the default aggregator to output the results in dependency order. (Examples such as `somepath(...) + dep(...)` will still output in lexicographical order as `somepath` is no longer at the top level).

This does not apply to `allpaths` since it does not output in a dependency order originally and hence there is no need to preserve any order (besides lexicographical).

PiperOrigin-RevId: 341612887