bazel runtime: fix cast error in BuildEventStreamer
commit 841ac61469ed61325acbcb134b6d372316f46540 changed expandSet to use NestedSetBuilder instead
of manipulating set internals directly. However, the builder
"inlines" new NestedSetBuilder().addTransitive(x).build() to just x,
which violates the postcondition of expandSet that all direct
elements are instances of class ExpandedArtifact.

(This can only happen if a non-leaf set has a single non-leaf
successor x, and one or more leaf successors all of which cause
visitArtifacts to be be a no-op, such as a middleman artifact.)

The fix is to handle the "inline" case specially, and apply
expandSet recursively to the resulting set.

Thanks to Janak for adding the test that eluded me.

BUG=159359614
PiperOrigin-RevId: 341684675