Propagate instrumented files for transitive sources of `android_library` and `android_binary`
RELNOTES: Propagate instrumented files for transitive sources of `android_library` and `android_binary`

BEGIN_PUBLIC

When Flutter applications are instrumented with `android_instrumentation_test`, non-Java source files cannot be instrumented. This is because these transitive source files are not propagated in the dependency tree in two places:

(simplified)

```
<flutter assets> <- android_library (1) <- android_binary <-instruments- android_binary (2) <- android_instrumentation_test
```

1. `android_library` does not propagate instrumented files from its dependencies
2. `android_binary` does not propagate instrumented files from the `instruments` attr

As a result, `android_instrumentation_test` cannot access the instrumented files of the `android_binary`, and the generated COVERAGE_MANIFEST will be empty. This CL addresses the above issues by fixing the chain to propagate the instrumented files.

This is not a problem for Java instrumented files in `android_instrumentation_test`, because of the lcov_merger has a special rule for Java source code – they do not need to be included in the COVERAGE_MANIFEST [1].

END_PUBLIC

PiperOrigin-RevId: 341718948