Automated rollback of commit 5137f36a1fcfc7ffec51a2c1588b0f90275e7a68.
*** Reason for rollback ***

We're rolling back this feature before cutting Bazel 4.0 LTS, as we can't merge the remaining part in time and are worried that the newly added patched JNI dependency will make it hard for distributions like Debian to accept.

Let's revisit this when we find a suitable pure Java version of Zstd.

*** Original change description ***

Add @com_github_luben_zstd_jni repo to WORKSPACE.

Partial import of #11968.

PiperOrigin-RevId: 342024383