Make runfiles-windows.cc not crash on dangling dummy_link
Since upgrading to bazel 3.6 we were experiencing similar issues to #11779. Occasionally a dangling dummy_link would exist (I think when the create runfiles process was terminated early) causing the next invocation to crash. This change handles the already exists error and continues execution in case the dummy_link is recognized as a symlink.

Closes #12428.

PiperOrigin-RevId: 342221243