Explicitly interrupt a cancelled strategy instead of relying on a misleading checkState
Future.cancel() is expected to return false if the future is done or had been
cancelled previously - a situation that often arises in stopBranch(). We therefore
want to remove the checkState for Future.cancel()'s return value.

However, it turned out that removing the checkState by itself would cause some
dynamic builds to fail: if the future of the strategy calling stopBranch() was
itself cancelled but not interrupted (a situation that sometimes arises), both
branches could attempt to write outputs. The IllegalStateException thrown by
checkState had the effect of interrupting the cancelled branch, allowing the build
to succeed most of the time, as long as the exception got caught somewhere (which
it often was).

We therefore want to replace the implicit effect of the misleading checkState with
an explicit check for a cancelled strategy and an explicit interruption.

RELNOTES: None.
PiperOrigin-RevId: 342340917