Throw a checked exception if there would be an illegal character in the output directory.
We do somewhat more granular checking so that we can error out with the most accurate possible message for the illegal character. This shouldn't have a big impact given the small number of BuildConfiguration/OutputDirectory objects created.

In some cases, this turns a crash into a silent modification of the top-level target, which seems equally bad: will not submit until this is resolved.

PiperOrigin-RevId: 342379102