Bump rules_cc to support llvm 11.0.0 clang-cl compiler on Windows
The clang-cl compiler version 11.0.0 number was missing a trailing whitespace that bazel rules_cc expected.

This commit picks up the most recent commit of rules_cc.

We would appreciate seeing this backported to 3.7.1 prior to the 4.0.0 release, since it represented a showstopper and has no side effects to users of earlier clang-cl compilers.

Signed-off-by: William A Rowe Jr <wrowe@vmware.com>

Closes #12458.

PiperOrigin-RevId: 342650689