Replace "javac" attribute with "tools" attribute in java_toolchain.
This is preparation for removal of "javac" attribute.

The main usecase for "javac" attribute is to use it in $location function during JVM parameters expansion. And this works when javac is added to tools attribute.

There is a starlark API to retrieve javac from the java toolchain, but without a known usecase. That API is not used internally and it is not tested.

PiperOrigin-RevId: 342894757