Use List instead of Map in Skyframe's intermediate eval results.
This CL is a no-op and only provides the necessary methods for Skyframe
evaluation with Lists instead of Maps. Actual application will be done in
subsequent CLs.

Constructing a Map<SkyKey, SkyValue> to store intermediate Skyframe evaluation result is wasteful since we don't make use of the look up capabilities of Maps, but rather just iterate over the key-value pairs. As an alternative, we can present the Skyframe evaluation result with a more space-efficient List<SkyValue>, then rely on the given order of the requested SkyKeys to iterate over it in various SkyFunctions.

Our benchmarks/profiles showed some reduction in GC and CPU time as a result of this switch.

RELNOTES: None
PiperOrigin-RevId: 342917514