Consolidate Interrupted failure detail codes into INTERRUPTED
We don't really need an enum for each command or call-site, that info can be put
in the error message. With this change we remove an excess of choices so that
it's easier for devs to do the right thing and to remove some cognitive burden
from the already tricky exercise of interrupt handling.

PiperOrigin-RevId: 342933703