Fix coverage tests so that they work on CI on Ubuntu 20
Progress on #12164. This fixes all the broken tests related to coverage and C++. There are other broken tests like Android that do not seem related at first glance.

The issue was that the system headers are different and this affects the coverage report generated for the example *.cc files we use in the test. The solution was to remove the use of these system headers, except for one place where it's important to test with at least one system header; there we just check for the absence of the system header in the generated report and the presence of the example file using this header but not its contents (which would be different from platform to platform)

RELNOTES:none
PiperOrigin-RevId: 343063978