Fix dangling reference to bazel.fish.
This was changed in https://github.com/bazelbuild/bazel/commit/20febacc0c02545684d3429fc21f28cf7ed8f461, but we forgot to update this reference. We only noticed this when trying to build the next release, because in the normal postsubmit the installer isn't built.

This CL also adds the release artifacts to the CI config so this cannot happen again.

PiperOrigin-RevId: 343119629