Starlark: Benchmarks --iterations <count> command line flag
Option to run a fixed number of iterations to use benchmarks with
an external comparison tool.

This is a follow-up to
[this comment](https://github.com/bazelbuild/bazel/pull/12498#issuecomment-729280929).

`Benchmarks --iterations 10000` will run 10000 iterations of a given
benchmark (or of the benchmarks if filter is not specified).

For example:

```
$ absh \
    -a "java -jar $HOME/Benchmarks_deploy-a.jar --filter bench_sort_small --iterations 10000000" \
    -b "java -jar $HOME/Benchmarks_deploy-b.jar --filter bench_sort_small --iterations 10000000"
...
A: N=53 r=3.992+-0.425 se=0.058
B: N=53 r=3.919+-0.384 se=0.053
B/A: 0.982
```

Note that currently `Benchmarks` parses all benchmark files (creating
unnecessary overhead), this will be fixed when new
`--filter <file>:<benchmark>` syntax is introduced.