Wire correct interrupt handling from ActionExecutedEvent
Generating command line arguments is an interruptible exercise. This might not
be ideal, but it is what it is. Instead ignoring-ish the interrupt, declare it
in all relevant places up the call stack.

This is primarily signature changes, with the exception of
FileTransport#asStreamProto which returns a null-valued future if the
computation is interrupted. Looking at the documentation and code, this should
be WAI.

PiperOrigin-RevId: 343157405