Respect interrupts coming from action key computation
This is because command line expansion is interruptible, a ship which has long
sailed. This is actually a net win, since things will now bail out sooner on
interrupt and have less time to corrupt state or otherwise do something wrong.

Most of this is adding exceptions to signatures. The main behavior changes are a
changed exception debug message (only if it's interrupted) and (rightfully)
bailing out of some async computation in action conflict code.

PiperOrigin-RevId: 343165180