Starlark: dict.pop should fail when frozen or when key is not hashable
... even if default value is provided.

[Starlark spec] says:

> `pop` fails if `key` is unhashable, or the dictionary is frozen or
> has active iterators.

It is not crystal clear how to interpet it.

My understanding is: if `dict` is frozen or key is not hashable,
`pop` must fail, even if default value is provided.

Similarly to how `list.extend([])` (extend with empty list) fails
if the list is frozen.

This PR should also make `dict.pop` a little faster.

[Starlark spec]: https://github.com/bazelbuild/starlark/blob/master/spec.md#dictpop