Delete seldom-used BuildViewTestCase helper method
useConfigurationMode() was a method for setting a value of --experimental_dynamic_configs that would be implicitly concatenated with the args of useConfiguration(). But it's clearer if we just make each test suite that needs it set it explicitly, or at least define its own helper for abstracting over that part of the test setup.

There are some tests that thought they were setting --experimental_dynamic_configs, but weren't in actuality. This was either due to the precedence of the default value of configsMode, or due to calling useConfigurationMode() without useConfiguration(). Some of these test cases fail now that they actually set the flag correctly. The failing ones are either deleted or else modified to comment out the flag (with a TODO).

Cleanup work toward #11437.

PiperOrigin-RevId: 343293996