Make execution phase SkyFunctions ready for --experimental_skyframe_eval_with_ordered_list.
Refactor ActionExecutionFunction and ArtifactNestedSetFunction to be able to pick the mode of Skyframe evaluation based on the value of --experimental_skyframe_eval_with_ordered_list. Reduction in GC is expected.

This is the first rollout, is flag-guarded for safety and requires a temporary increase in the size of the codebase.

RELNOTES: None
PiperOrigin-RevId: 343297169