starlark: implement "flat frames" optimization
The resolver now statically maps each local variable to a small
integer, and the set of local bindings (including from
comprehensions) is accumulated into each Resolver.Function.

The evaluator uses an array instead of a LinkedHashMap to
represent the environment, and no longer needs a hack to
save/restore the values of comprehensions.

Also, added a test of the resolver.

PiperOrigin-RevId: 343529575