starlark resolver: implement "flat globals" optimization
This change reworks the Resolver.Module interface (which abstracts
eval.Module) so that each global is statically mapped to a small integer,
so that they can be represented during execution as a flat array,
not a hash table, similar to the treatment of locals in CL 343371120.

It also resolves PREDECLARED and UNIVERSE references distinctly, so that
we needn't try both maps during execution. And it sets the stage for
those maps to be treated as flat arrays too.

Furthermore, it materializes for the first time a "file-local block",
which will be used to hold bindings created by load statements, when
we make loads bind locally and eliminate the getExportedGlobals hack.

The Resolver.Module interface no longer requires the implementation to
enumerate all the defined names; names are looked up on demand as
references are encountered in the program during resolution, and
Bindings are created for them. This reduces unnecessary allocation and
copying when a small program is executed in a big environment.
The resolver still provides spelling suggestions over all available
names; it does this by having the Module implementation provide
the set of top-level candidates lazily: only when resolve() fails.

PiperOrigin-RevId: 343776166