Removing javac from JavaToolchain provider and Starlark API.
Nothing is reading javac from JavaToolchain provider. There is a Starlark interface for it, but no users. It is hard to imagine a usecase where somebody would actually need javac.jar.
I move to bluntly remove it from the Starlark API, without guarding it with a flag or documenting it was once there.

RELNOTES[INC]: Removing java_common.javac_jar Starlark call.

PiperOrigin-RevId: 343797870