Starlark: do not show .. prompt in REPL on complete input
```
$ bazel run //src/main/java/net/starlark/java/cmd:starlark
...
Welcome to Starlark (java.starlark.net)
>> x = 1
>> y = [
..   "a",
..   "b",
.. ]
>>
>> def
..
<stdin>:1:4: syntax error at 'newline': expected identifier
>>
```

Ideally, REPL lexer/parser should be incremental: when it reads
line, it should keep it state and continue parsing the next line.
But it is hard to implement and this PR proposes simpler solution:
REPL read lines until either lines produce valid statements or empty
line is given. This overhead should not be an issue for REPL.