Allow using embedded tools in sandboxed spawn runners.
When using sandboxed runners, the only supported `VirtualActionInput` types are
empty inputs and param files. Add support for `BinTools.PathActionInput` so
that we can we can use embedded tools in sandboxed runners. Make sure to write
them as executable to the sandbox -- contrary to param files, embedded tools
can be run.

PiperOrigin-RevId: 344438370