Fix coverage tests so that they work on CI on Ubuntu 20
Progress on #12164. Fixes remaining coverage test. The issue was caused by
Ubuntu 20 being the first distribution that we have on CI to use gcc 9. From
gcc 9 and above the intermediate format used by gcov is different and relies on
json zipped files. I adapted the test to work with version 9 and higher and
left the code that tests gcc versions 7 and 8 for now, but will be removed
eventually.

RELNOTES:none
PAIR=cmita
PiperOrigin-RevId: 344502710