Correct documentation for filegroup.data
It's true that filegroup.data ends up in the consumer's runfiles when consumed via the data attribute, but that's also true for filegroup.srcs. The point of filegroup.data is that it provides stuff that's used by whatever is in filegroup.srcs at runtime, and thus should be included in runfiles whether or not filegroup.srcs itself is supposed to end up in runfiles.

For example, if filegroup.srcs contains source code, filegroup.data might contain things that the code expects to find in runfiles when it is executed. If such a filegroup target is depended on via a native rule's srcs attribute, filegroup.data ends up in the consumer's runfiles, but filegroup.srcs does not. (If such a filegroup target is depended on via a native rule's data attribute, both filegroup.srcs and filegroup.data end up in the consumer's runfiles.)

Some of the wording is made a little more direct, e.g. "will be considered belonging to" -> "is added to".

Suggest including srcs in the attributes runfiles are merged from in Starlark rules, so that including filegroups in srcs works as expected if those filegroups have data. This is consistent with how native rules behave.

RENOTES: None.
PiperOrigin-RevId: 344796745