Filter coverage output to just the files in the source file manifest
Coverage output should be filtered to just the sources listed in the manifest because those are the sources of targets matched by --instrumentation_filter (excluding test targets unless --instrument_test_targets). Coverage output should be restricted to those files according to the documentation for those flags.

This is a particular concern for C++ because C++ source files can include header files. That is, header files are source files of the target which directly includes them, but unlike most source files, code in header files may also be compiled by that target's consumers. Thus, CppCompile actions enable coverage instrumentation if either the target or any of the target's direct dependencies are matched by --instrumentation_filter. The resulting coverage output needs to be filtered to just the specific targets that are matched.

However, C++ is not the only language which has that property. It's also possible for coverage implementations to generate overly-broad coverage data for other reasons, or as an oversight. In those cases, the semantics of --instrumentation_filter should still be enforced.

This adds "jars" and "runtime_jars" to dependency attributes in the InstrumentationSpec for Java rules in order to preserve behavior for some of the tests. That test case doesn't seem like a usual way of structuring that dependency, if you have the source files checked in (as opposed to just the jarfile), it makes more sense to use the java_library target instead of using the jarfile output. But "jars" and "resource_jars" are in fact runtime dependencies, so listing them is at least technically correct.

RELNOTES: Filter all (instead of just C++) source files for coverage output according to --instrumentation_filter and --instrument_test_targets.
PiperOrigin-RevId: 344797828